chatbot_openai/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
Flask==3.1.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.19
langchain-core==0.3.40
langchain-openai==0.3.7
langchain-text-splitters==0.3.6
langsmith==0.3.11
MarkupSafe==3.0.2
multidict==6.1.0
numpy==1.26.4
openai==1.65.2
orjson==3.10.15
packaging==24.2
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
simple-websocket==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.38
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.18.3
zstandard==0.23.0
